backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
pydantic
google-generativeai
python-dotenv
sqlalchemy
aiosqlite
alembic